+++ templates/slots/Services/Repository/RepositoryObject/sql/dbupdate.php
@@ -1,5 +1,5 @@
 <#1>
 <?php
 \__NAMESPACE__\Config\Config::updateDB();
-\__NAMESPACE__\Object\Obj::updateDB();
+\__NAMESPACE__\ObjectSettings\ObjectSettings::updateDB();
 ?>