+++ templates/slots/Services/UIComponent/UserInterfaceHook/src/GUI/class.GUI.php
@@ -10,8 +10,7 @@ use srag\DIC\__PLUGIN_NAME__\DICTrait;
  *
  * @package           __NAMESPACE__\GUI
  *
- * @author            studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
- * @author            __RESPONSIBLE_NAME__ <__RESPONSIBLE_EMAIL__>
+ * __AUTHOR_COMMENT__
  *
  * @ilCtrl_isCalledBy __NAMESPACE__\GUI\GUI: ilUIPluginRouterGUI
  */