+++ templates/slots/base/src/Config/ConfigFormGUI.php
@@ -12,8 +12,7 @@ use srag\CustomInputGUIs\__PLUGIN_NAME__\PropertyFormGUI\ConfigPropertyFormGUI;
  *
  * @package __NAMESPACE__\Config
  *
- * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
- * @author  __RESPONSIBLE_NAME__ <__RESPONSIBLE_EMAIL__>
+ * __AUTHOR_COMMENT__
  */
 class ConfigFormGUI extends ConfigPropertyFormGUI
 {